+++ tests/advection/polar_foot_finder.cpp
@@ -9,11 +9,11 @@
 #include "czarny_to_cartesian.hpp"
 #include "ddc_aliases.hpp"
 #include "geometry_pseudo_cartesian.hpp"
+#include "interpolation_polar_foot_finder.hpp"
 #include "mesh_builder.hpp"
 #include "r_theta_test_cases.hpp"
 #include "rk4.hpp"
 #include "species_info.hpp"
-#include "spline_polar_foot_finder.hpp"
 #include "vector_field.hpp"
 #include "vector_field_mem.hpp"
 #include "vector_index_tools.hpp"
