+++ tests/geometryRTheta/advection_rtheta/advection_all_tests.cpp
@@ -25,6 +25,7 @@
 #include "euler.hpp"
 #include "geometry_r_theta.hpp"
 #include "input.hpp"
+#include "interpolation_polar_foot_finder.hpp"
 #include "math_tools.hpp"
 #include "mesh_builder.hpp"
 #include "paraconfpp.hpp"
@@ -34,7 +35,6 @@
 #include "rk3.hpp"
 #include "rk4.hpp"
 #include "spline_definitions_r_theta.hpp"
-#include "spline_polar_foot_finder.hpp"
 #include "vector_field.hpp"
 #include "vector_field_mem.hpp"
