+++ tests/geometryRTheta/advection_rtheta/advection_selected_test.cpp
@@ -24,6 +24,7 @@
 #include "geometry_pseudo_cartesian.hpp"
 #include "geometry_r_theta.hpp"
 #include "input.hpp"
+#include "interpolation_polar_foot_finder.hpp"
 #include "itimestepper.hpp"
 #include "math_tools.hpp"
 #include "mesh_builder.hpp"
@@ -34,7 +35,6 @@
 #include "rk3.hpp"
 #include "rk4.hpp"
 #include "spline_definitions_r_theta.hpp"
-#include "spline_polar_foot_finder.hpp"
