+++ tests/geometryRTheta/advection_rtheta/advection_without_Opoint.cpp
@@ -28,6 +28,7 @@
 #include "discrete_poloidal_cs_spline_mapping_builder.hpp"
 #include "euler.hpp"
 #include "geometry_r_theta.hpp"
+#include "interpolation_polar_foot_finder.hpp"
 #include "inverse_jacobian_matrix.hpp"
 #include "l_norm_tools.hpp"
 #include "mesh_builder.hpp"
@@ -36,7 +37,6 @@
 #include "rk4.hpp"
 #include "simulation_utils_tools.hpp"
 #include "spline_definitions_r_theta.hpp"
-#include "spline_polar_foot_finder.hpp"
 #include "spline_quadrature.hpp"
 #include "trapezoid_quadrature.hpp"
